[fix][node]fix instance and node detail

+++ dingo_command/api/node.py
@@ -14,6 +14,7 @@ node_service = NodeService()
 async def list_nodes(cluster_id: str = Query(None, description="集群id"),
                      cluster_name: str = Query(None, description="集群名称"),
                      type: str = Query(None, description="节点类型"),
+                     status: str = Query(None, description="status状态"),
                      page: int = Query(1, description="页码"),
                      page_size: int = Query(10, description="页数量大小"),
                      sort_keys: str = Query(None, description="排序字段"),
@@ -26,7 +27,8 @@ async def list_nodes(cluster_id: str = Query(None, description="集群id"),
             query_params['cluster_name'] = cluster_name
         if type:
             query_params['type'] = type
-        query_params = {}
+        if status:
+            query_params['status'] = status
         # 查询条件组装
         if cluster_id:
             query_params['cluster_id'] = cluster_id
@@ -55,18 +57,19 @@ async def get_node(node_id: str):
 
 @router.post("/node", summary="扩容节点", description="扩容节点")
 async def create_node(cluster: ScaleNodeObject):
+    # 先检查下是否有正在处于扩容的状态，如果是就直接返回
+    cluster_service = ClusterService()
+    cluster_info = cluster_service.get_cluster(cluster.id)
+    if not cluster_info:
+        raise HTTPException(status_code=400, detail="the cluster does not exist, please check")
+    if cluster_info.status == "creating":
+        raise HTTPException(status_code=400, detail="the cluster is creating, please wait")
+    if cluster_info.status == "scaling":
+        raise HTTPException(status_code=400, detail="the cluster is scaling, please wait")
     try:
-        # 先检查下是否有正在处于扩容的状态，如果是就直接返回
-        cluster_service = ClusterService()
-        cluster_info = cluster_service.get_cluster(cluster.id)
-        if not cluster_info:
-            raise HTTPException(status_code=400, detail="the cluster does not exist, please check")
-        # if result.status == "scaling":
-        #     raise HTTPException(status_code=400, detail="the cluster is scaling, please wait")
-
         # 创建节点（扩容节点）
         result = node_service.create_node(cluster_info, cluster)
-        return result
+        return {"data": result}
     except Fail as e:
         raise HTTPException(status_code=400, detail=e.error_message)
     except Exception as e:
@@ -77,27 +80,26 @@ async def create_node(cluster: ScaleNodeObject):
 
 @router.delete("/node/{node_id}", summary="缩容节点", description="缩容节点")
 async def delete_node(node_id: str):
-    try:
-        data = node_service.get_node(node_id)
-        if data.get("data") is not None:
-            node = data.get("data")
-            # 先检查下是否有正在处于缩容的状态，如果是就直接返回
-            cluster_service = ClusterService()
-            result = cluster_service.get_cluster(node.cluster_id)
-            if not result:
-                raise HTTPException(status_code=400, detail="the cluster does not exist, please check")
-            # if result.status == "removing":
-            #     raise HTTPException(status_code=400, detail="the cluster is removing, please wait")
-
+    data = node_service.get_node(node_id)
+    if data.get("data") is not None:
+        node = data.get("data")
+        # 先检查下是否有正在处于缩容的状态，如果是就直接返回
+        cluster_service = ClusterService()
+        result = cluster_service.get_cluster(node.cluster_id)
+        if not result:
+            raise HTTPException(status_code=400, detail="the cluster does not exist, please check")
+        if result.status == "creating":
+            raise HTTPException(status_code=400, detail="the cluster is creating, please wait")
+        try:
             # 缩容某些节点
             result = node_service.delete_node(node)
             if result is not None:
                 return {"data": result}
-        else:
-            return {"data": None}
-    except Fail as e:
-        raise HTTPException(status_code=400, detail=e.error_message)
-    except Exception as e:
-        import traceback
-        traceback.print_exc()
-        raise HTTPException(status_code=400, detail="delete node error")
+        except Fail as e:
+            raise HTTPException(status_code=400, detail=e.error_message)
+        except Exception as e:
+            import traceback
+            traceback.print_exc()
+            raise HTTPException(status_code=400, detail="delete node error")
+    else:
+        raise HTTPException(status_code=400, detail="can not find the node, please check")
